LoginAction refactor
