+++ src/main/java/UseCases/LoginAction/LoginAction.java
@@ -1,7 +1,11 @@
-package UseCases;
+package UseCases.LoginAction;
 
 import DataBase.*;
 import Entities.*;
+/**
+ * A class that implements LoginActionUserCredentials calls upon DataBase and User in order to implement a login
+ * functionality
+ */
 public class LoginAction {
     private String inputtedUsername;
     private String inputtedPassword;
